Rename tcg_dice to riot
