+++ scheme/riot/Makefile
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 .DEFAULT_GOAL := test
 
-GOPKG := github.com/veraison/services/scheme/tcg-dice
+GOPKG := github.com/veraison/services/scheme/riot
 SRCS := $(wildcard *.go)
 
 SUBDIR += plugin
