+++ scheme/riot/evidence_handler_test.go
@@ -1,7 +1,7 @@
 // Copyright 2021-2023 Contributors to the Veraison project.
 // SPDX-License-Identifier: Apache-2.0
 
-package tcg_dice
+package riot
 
 import (
 	"os"
