+++ scheme/riot/plugin/combined/main.go
@@ -5,7 +5,7 @@ package main
 import (
 	"github.com/veraison/services/handler"
 	"github.com/veraison/services/plugin"
-	scheme "github.com/veraison/services/scheme/tcg-dice"
+	scheme "github.com/veraison/services/scheme/riot"
 )
 
 func main() {
