+++ scheme/riot/scheme.go
@@ -1,8 +1,8 @@
 // Copyright 2023 Contributors to the Veraison project.
 // SPDX-License-Identifier: Apache-2.0
-package tcg_dice
+package riot
 
-const SchemeName = "TCG_DICE"
+const SchemeName = "riot"
 
 var EvidenceMediaTypes = []string{
 	"application/pem-certificate-chain",
